Farkas Tibor Custom Page Template
This project involved creating a fully custom page template for a one-page website. The client initially chose a theme, but it was later converted to a custom page template from scratch using a child theme. I worked on this project from start to finish, ensuring a fully custom template integrated with Advanced Custom Fields (ACF) and CMB2 for repeater functionality.